Button - fix double submit issues
