+++ src/components/button/_test-template.njk
@@ -1,5 +1,5 @@
 {% from "components/button/_macro.njk" import onsButton %}
 
-<form class="js-patternlib-form" onsubmit="return false;">
+<form onsubmit="return false;">
     {{ onsButton(params) }}
 </form>
