+++ src/components/button/examples/button-loader/index.njk
@@ -1,13 +1,12 @@
 {% from "components/button/_macro.njk" import onsButton %}
-<form class="js-patternlib-form" onsubmit="return false;">
+<form onsubmit="return false;">
     <div class="btn-group">
-      {{
-          onsButton({
-              "text": 'Submit',
-              "id": 'submit-btn',
-              "classes": 'btn-group__btn',
-              "submitType": "loader"
-          })
-      }}
+    {{
+        onsButton({
+            "text": 'Submit',
+            "classes": 'btn-group__btn',
+            "submitType": "loader"
+        })
+    }}
     </div>
 </form>
