--- src/tests/spec/button/timer-button.spec.js
@@ -1,66 +1,0 @@
-import { awaitPolyfills } from 'js/polyfills/await-polyfills';
-import template from 'components/button/_test-template.njk';
-import SubmitButton from 'components/button/button';
-
-describe('Function: Timer Button ', function() {
-  let wrapper, buttonElement;
-
-  let params = {
-    id: 'button',
-    text: 'Submit',
-    submitType: 'timer',
-  };
-
-  before(() => awaitPolyfills);
-
-  beforeEach(() => {
-    const html = template.render({ params });
-
-    wrapper = document.createElement('div');
-    wrapper.innerHTML = html;
-    document.body.appendChild(wrapper);
-
-    buttonElement = document.getElementById(params.id);
-  });
-
-  afterEach(() => {
-    if (wrapper) {
-      wrapper.remove();
-    }
-  });
-
-  describe('Before the button is initialised', () => {
-    it('Button should have relevant classes', () => {
-      expect(buttonElement.classList.contains('js-timer')).to.be.true;
-      expect(buttonElement.classList.contains('js-submit-btn')).to.be.true;
-    });
-  });
-
-  describe('Once the button is initialised', () => {
-    beforeEach(() => {
-      new SubmitButton(buttonElement, params.submitType);
-    });
-
-    it('Button disabled attribute should not be set', () => {
-      expect(buttonElement.getAttribute('disabled')).to.not.exist;
-    });
-
-    describe('and the button is clicked', () => {
-      beforeEach(() => {
-        buttonElement.click();
-      });
-
-      it('Button should be disabled', () => {
-        expect(buttonElement.getAttribute('disabled')).to.equal('true');
-      });
-    });
-
-    describe('one second after the button is clicked', () => {
-      it('the button disabled attribute should not be set', () => {
-        setTimeout(() => {
-          expect(buttonElement.getAttribute('disabled')).to.not.exist;
-        }, 1000);
-      });
-    });
-  });
-});
